Ice 16? I also noticed today that my logs didnt commit
Whoops

+++ src/server/controllers/animals/delete.js
@@ -3,16 +3,19 @@ import AnimalService from "../../services/AnimalService.js";
 
 const rules = checkSchema({
     animalId: {
-        in: ['params'],
             errorMessage: 'Invalid animal ID',
         },
             custom: {
                 options: async (value) =>{
-                    if (!await AnimalService.deleteAnimal(value)) {
-                        throw new Error('That animal does not exists');
+                if (!isValidObjectId (value)) {
+                    throw new Error('Invalid ID');
                 }
-            }
-        },
+                if (!await AnimalService.deleteAnimal(value)) {
+                    throw new Error('That animal does not exists');
+                }
+            },
+        in: 'params',
+        }
 });
 
 
@@ -20,7 +23,7 @@ const handle = async (req, res, next) => {
     try {
         const { animalId } = req.params;
         const animal = await AnimalService.deleteAnimal(animalId);
-        res.json(`${animal} has been deleted :( `);
+        res.json(`${animal.name} has been deleted :( `);
 
     } catch (error) {
         next(error);
